Move FormResults component into shared Admin component folder
