+++ front/app/components/admin/FormResults/FormResultsQuestion/MappingQuestions/LineLocationQuestion/index.tsx
