+++ front/app/components/admin/FormResults/FormResultsQuestion/MappingQuestions/PointLocationQuestion/index.tsx
