+++ front/app/components/admin/FormResults/FormResultsQuestion/MappingQuestions/PolygonLocationQuestion/index.tsx
