+++ front/app/components/admin/FormResults/FormResultsQuestion/MatrixQuestion/utils.tsx
