+++ front/app/components/admin/FormResults/FormResultsQuestion/NumberQuestion/NumberResponses.tsx
