+++ front/app/components/admin/FormResults/FormResultsQuestion/NumberQuestion/index.tsx
