+++ front/app/components/admin/FormResults/FormResultsQuestion/RankingQuestion/DetailedRankView.tsx
