+++ front/app/components/admin/FormResults/FormResultsQuestion/RankingQuestion/DetailedViewButton.tsx
