+++ front/app/components/admin/FormResults/FormResultsQuestion/RankingQuestion/index.tsx
