+++ front/app/components/admin/FormResults/FormResultsQuestion/RankingQuestion/messages.ts
