+++ front/app/components/admin/FormResults/FormResultsQuestion/RankingQuestion/utils.tsx
