+++ front/app/components/admin/FormResults/FormResultsQuestion/TextQuestion/Analysis/AnalysisInsights.tsx
