+++ front/app/components/admin/FormResults/FormResultsQuestion/TextQuestion/Analysis/Question.tsx
