+++ front/app/components/admin/FormResults/FormResultsQuestion/TextQuestion/Analysis/Summary.tsx
