+++ front/app/components/admin/FormResults/FormResultsQuestion/TextQuestion/Analysis/index.tsx
