+++ front/app/components/admin/FormResults/FormResultsQuestion/TextQuestion/Analysis/utils.ts
