+++ front/app/components/admin/FormResults/FormResultsQuestion/components/FormResultsQuestionValue.tsx
