+++ front/app/components/admin/FormResults/FormResultsQuestion/utils.tsx
