+++ front/app/components/admin/FormResults/LogicIcon.tsx
