+++ front/app/components/admin/Graphs/SurveyBars/SurveyBarsHorizontal.tsx
@@ -10,7 +10,7 @@ import {
 
 import useLocalize from 'hooks/useLocalize';
 
-import LogicIcon from 'containers/Admin/projects/project/nativeSurvey/FormResults/LogicIcon';
+import LogicIcon from 'components/admin/FormResults/LogicIcon';
 
 import { useIntl } from 'utils/cl-intl';
