+++ front/app/containers/Admin/communityMonitor/components/LiveMonitor/index.tsx
@@ -4,7 +4,7 @@ import { Box, colors, Title } from '@citizenlab/cl2-component-library';
 
 import useCommunityMonitorProject from 'api/community_monitor/useCommunityMonitorProject';
 
-import FormResults from 'containers/Admin/projects/project/nativeSurvey/FormResults';
+import FormResults from 'components/admin/FormResults';
 
 import { FormattedMessage } from 'utils/cl-intl';
