+++ front/app/containers/Admin/reporting/components/ReportBuilder/Widgets/SurveyQuestionResultWidget/Question/MapQuestions/PolygonLocationQuestion.tsx
@@ -5,7 +5,7 @@ import { Spinner } from '@citizenlab/cl2-component-library';
 import useMapConfigById from 'api/map_config/useMapConfigById';
 import useProjectMapConfig from 'api/map_config/useProjectMapConfig';
 
-import surveyResultMessages from 'containers/Admin/projects/project/nativeSurvey/FormResults/messages';
+import surveyResultMessages from 'components/admin/FormResults/messages';
 
 import PolygonMap from 'components/admin/Graphs/PolygonMap';
