+++ front/app/containers/Admin/reporting/components/ReportBuilder/Widgets/SurveyQuestionResultWidget/Question/index.tsx
@@ -7,9 +7,9 @@ import { GroupMode } from 'api/graph_data_units/requestTypes';
 
 import useLocalize from 'hooks/useLocalize';
 
-import InputType from 'containers/Admin/projects/project/nativeSurvey/FormResults/FormResultsQuestion/InputType';
-import MatrixQuestion from 'containers/Admin/projects/project/nativeSurvey/FormResults/FormResultsQuestion/MatrixQuestion';
-import RankingQuestion from 'containers/Admin/projects/project/nativeSurvey/FormResults/FormResultsQuestion/RankingQuestion';
+import InputType from 'components/admin/FormResults/FormResultsQuestion/InputType';
+import MatrixQuestion from 'components/admin/FormResults/FormResultsQuestion/MatrixQuestion';
+import RankingQuestion from 'components/admin/FormResults/FormResultsQuestion/RankingQuestion';
 
 import Legend from 'components/admin/Graphs/Legend';
 import { DEFAULT_CATEGORICAL_COLORS } from 'components/admin/Graphs/styling';
