+++ front/app/containers/Admin/reporting/components/ReportBuilder/Widgets/SurveyQuestionResultWidget/Settings/index.tsx
@@ -15,7 +15,7 @@ import { ICustomFields } from 'api/custom_fields/types';
 import useRawCustomFields from 'api/custom_fields/useRawCustomFields';
 import { GroupMode } from 'api/graph_data_units/requestTypes';
 
-import HeatmapTooltipContent from 'containers/Admin/projects/project/nativeSurvey/FormResults/FormResultsQuestion/MappingQuestions/PointLocationQuestion/HeatmapTooltipContent';
+import HeatmapTooltipContent from 'components/admin/FormResults/FormResultsQuestion/MappingQuestions/PointLocationQuestion/HeatmapTooltipContent';
 import nativeSurveyMessages from 'containers/Admin/projects/project/nativeSurvey/messages';
 
 import PhaseFilter from 'components/UI/PhaseFilter';
